refactor: extract release planning skip policy
