+++ src/core/release/mod.rs
@@ -7,6 +7,7 @@ mod pipeline;
 mod pipeline_capabilities;
 mod pipeline_summary;
 mod plan_steps;
+mod planning_policy;
 mod types;
 mod utils;
 pub mod version;
